Moved object meet / item utils.
