+++ src/engine/core/managers/debug/DebugManager.ts
@@ -5,10 +5,10 @@ import { AbstractManager } from "@/engine/core/managers/base/AbstractManager";
 import { StalkerStateManager } from "@/engine/core/objects/ai/state/StalkerStateManager";
 import { EActionId, EStateActionId } from "@/engine/core/objects/ai/types";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { getObjectActiveWeaponSlot } from "@/engine/core/utils/object";
 import { getNumberRelationBetweenCommunities } from "@/engine/core/utils/relation";
 import { gameTimeToString } from "@/engine/core/utils/time";
 import { toJSON } from "@/engine/core/utils/transform/json";
+import { getObjectActiveWeaponSlot } from "@/engine/core/utils/weapon";
 import { stalkerCommunities, TCommunity } from "@/engine/lib/constants/communities";
 import { NIL } from "@/engine/lib/constants/words";
 import { ActionPlanner, ClientObject, ESchemeType, Optional, TLabel, TName, TNumberId } from "@/engine/lib/types";
