+++ src/engine/core/managers/drop/DropManager.ts
@@ -8,8 +8,8 @@ import { abort } from "@/engine/core/utils/assertion";
 import { isArtefact, isGrenade, isWeapon } from "@/engine/core/utils/class_ids";
 import { getObjectCommunity } from "@/engine/core/utils/community";
 import { parseNumbersList, parseStringsList } from "@/engine/core/utils/ini";
+import { setItemCondition } from "@/engine/core/utils/item";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { setItemCondition } from "@/engine/core/utils/object";
 import { isAmmoSection, isExcludedFromLootDropItemSection, isLootableItemSection } from "@/engine/core/utils/section";
 import { spawnItemsForObject } from "@/engine/core/utils/spawn";
 import { logicsConfig } from "@/engine/lib/configs/LogicsConfig";
