+++ src/engine/core/objects/ai/state/weapon/ActionWeaponDrop.ts
@@ -2,8 +2,8 @@ import { action_base, LuabindClass, object } from "xray16";
 
 import { StalkerStateManager } from "@/engine/core/objects/ai/state/StalkerStateManager";
 import { isStrappableWeapon } from "@/engine/core/utils/class_ids";
+import { setItemCondition } from "@/engine/core/utils/item";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { setItemCondition } from "@/engine/core/utils/object";
 import { getObjectWeaponForAnimationState } from "@/engine/core/utils/weapon";
 import { logicsConfig } from "@/engine/lib/configs/LogicsConfig";
 import { ClientObject, Optional } from "@/engine/lib/types";
