+++ src/engine/core/schemes/stalker/meet/utils/meet_handling.ts
@@ -1,19 +1,15 @@
 import { registry } from "@/engine/core/database";
 import { GlobalSoundManager } from "@/engine/core/managers/sounds/GlobalSoundManager";
+import { ISchemeAbuseState } from "@/engine/core/schemes/stalker/abuse";
 import { ISchemeMeetState } from "@/engine/core/schemes/stalker/meet";
 import { MeetManager } from "@/engine/core/schemes/stalker/meet/MeetManager";
 import { ISchemeWoundedState } from "@/engine/core/schemes/stalker/wounded";
 import { pickSectionFromCondList } from "@/engine/core/utils/ini";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import {
-  addObjectAbuse,
-  isObjectHelpingWounded,
-  isObjectSearchingCorpse,
-  isObjectWounded,
-} from "@/engine/core/utils/object";
+import { isObjectHelpingWounded, isObjectSearchingCorpse, isObjectWounded } from "@/engine/core/utils/object";
 import { getObjectsRelationSafe } from "@/engine/core/utils/relation";
 import { FALSE, NIL, TRUE } from "@/engine/lib/constants/words";
-import { ClientObject, EClientObjectRelation, EScheme, Optional, TName } from "@/engine/lib/types";
+import { ClientObject, EClientObjectRelation, EScheme, Optional, TCount, TName } from "@/engine/lib/types";
 
 const logger: LuaLogger = new LuaLogger($filename, { file: "meet" });
 
@@ -91,5 +87,44 @@ export function activateMeetWithObject(object: ClientObject): void {
     getObjectsRelationSafe(object, actor) !== EClientObjectRelation.ENEMY
   ) {
     addObjectAbuse(object, 1);
+  }
+}
+
+/**
+ * Increment abuse for object.
+ *
+ * @param object - target client object
+ * @param value - count of abuse to add
+ */
+export function addObjectAbuse(object: ClientObject, value: TCount): void {
+  const abuseState: Optional<ISchemeAbuseState> = registry.objects.get(object.id())[EScheme.ABUSE] as ISchemeAbuseState;
+
+  abuseState?.abuseManager.addAbuse(value);
+}
+
+/**
+ * Clear abuse state for object.
+ *
+ * @param object - target client object
+ */
+export function clearObjectAbuse(object: ClientObject): void {
+  const state: Optional<ISchemeAbuseState> = registry.objects.get(object.id())[EScheme.ABUSE] as ISchemeAbuseState;
+
+  state?.abuseManager.clearAbuse();
+}
+
+/**
+ * Set object abuse state.
+ *
+ * @param object - target client object
+ * @param isEnabled - whether object abuse state should be enabled
+ */
+export function setObjectAbuseState(object: ClientObject, isEnabled: boolean): void {
+  const state: Optional<ISchemeAbuseState> = registry.objects.get(object.id())[EScheme.ABUSE] as ISchemeAbuseState;
+
+  if (isEnabled) {
+    state?.abuseManager.enableAbuse();
+  } else {
+    state?.abuseManager.disableAbuse();
   }
 }
