+++ src/engine/core/utils/weapon.ts
@@ -6,6 +6,22 @@ import { isWeapon } from "@/engine/core/utils/class_ids";
 import { ClientObject, Optional, TIndex } from "@/engine/lib/types";
 
 /**
+ * Get active weapon slot of an object for animating.
+ *
+ * @param object - target client object to check
+ * @returns active weapon slot index
+ */
+export function getObjectActiveWeaponSlot(object: ClientObject): TIndex {
+  const weapon: Optional<ClientObject> = object.active_item();
+
+  if (weapon === null || object.weapon_strapped()) {
+    return 0;
+  }
+
+  return weapon.animation_slot();
+}
+
+/**
  * todo;
  */
 export function isObjectWeaponLocked(object: ClientObject): boolean {
